requirements.txt
pyyml==0.0.2
websockets==14.2
opuslib_next==1.1.2
numpy==1.26.4
pydub==0.25.1
google-generativeai==0.8.4
httpx==0.27.2
aiohttp==3.9.3
aiohttp_cors==0.7.0
ormsgpack==1.7.0
ruamel.yaml==0.18.10
loguru==0.7.3
requests==2.32.3
cozepy==0.12.0
modelscope==1.23.2
sherpa_onnx==1.12.0
mcp==1.8.1
cnlunar==0.2.0
PySocks==1.7.1
dashscope==1.23.1
chardet==5.2.0
aioconsole==0.8.1
markitdown==0.1.1
mcp-proxy==0.6.0
PyJWT==2.8.0
psutil==7.0.0
portalocker==2.10.1